% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LRs.R
\name{compute_LRs}
\alias{compute_LRs}
\title{Compute Likelihood Ratios (LRs) for Matching Personal Characteristics}
\usage{
compute_LRs(data, h, s, y, eh, es, ey)
}
\arguments{
\item{data}{A data.frame containing the characteristics of the UPs.}

\item{h}{An integer representing the hair color of the MP.}

\item{s}{An integer representing the skin color of the MP.}

\item{y}{An integer representing the eye color of the MP.}

\item{eh}{A numeric value representing the error rate for observing hair color.}

\item{es}{A numeric value representing the error rate for observing skin color.}

\item{ey}{A numeric value representing the error rate for observing eye color.}
}
\value{
A numeric vector containing the LRs for each individual in the dataset, calculated
based on the probability of observing the given combination of characteristics in the MP
and comparing it with each UP in the dataset.
}
\description{
This function calculates the Likelihood Ratios (LRs) for each individual in a dataset
based on the comparison of hair color, skin color, and eye color between a Missing Person (MP)
and the Unidentified Persons (UPs) in the dataset. It also accounts for error rates
in the observation of each characteristic.
}
\examples{
compute_LRs(data, 1, 1, 1, 0.01, 0.01, 0.01) # Calculate LRs for a given set of MP characteristics and error rates.
}
