% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRs.R
\name{plotLRs}
\alias{plotLRs}
\title{Plot Distribution of Likelihood Ratios}
\usage{
plotLRs(LRs)
}
\arguments{
\item{LRs}{A numeric vector containing the Likelihood Ratios computed by the compute_LRs function.}
}
\description{
This function creates a density plot for the distribution of Likelihood Ratios (LRs)
calculated for a dataset.
}
\examples{
# Assuming 'LRs' is the output from compute_LRs function
plotLRs(LRs)
}
